Breakout LineDrawable
