+++ src/Svg.Skia/Svg/LineDrawable.cs
@@ -1,0 +1,82 @@
+﻿#if !USE_PICTURE
+using SkiaSharp;
+using CropRect = SkiaSharp.SKImageFilter.CropRect;
+#endif
+#if USE_PICTURE
+using SKRect = Svg.Picture.Rect;
+#endif
+
+namespace Svg.Skia
+{
+    internal sealed class LineDrawable : DrawablePath
+    {
+        private LineDrawable()
+            : base()
+        {
+        }
+
+        public static LineDrawable Create(SvgLine svgLine, SKRect skOwnerBounds, DrawableBase? parent, Attributes ignoreAttributes = Attributes.None)
+        {
+            var drawable = new LineDrawable
+            {
+                Element = svgLine,
+                Parent = parent,
+                IgnoreAttributes = ignoreAttributes
+            };
+
+            drawable.IsDrawable = drawable.CanDraw(svgLine, drawable.IgnoreAttributes) && drawable.HasFeatures(svgLine, drawable.IgnoreAttributes);
+
+            if (!drawable.IsDrawable)
+            {
+                return drawable;
+            }
+
+            drawable.Path = svgLine.ToSKPath(svgLine.FillRule, skOwnerBounds, drawable.Disposable);
+            if (drawable.Path is null || drawable.Path.IsEmpty)
+            {
+                drawable.IsDrawable = false;
+                return drawable;
+            }
+
+            drawable.IsAntialias = SvgExtensions.IsAntialias(svgLine);
+
+            drawable.TransformedBounds = drawable.Path.Bounds;
+
+            drawable.Transform = SvgExtensions.ToSKMatrix(svgLine.Transforms);
+
+            bool canDrawFill = true;
+            bool canDrawStroke = true;
+
+            if (SvgExtensions.IsValidFill(svgLine))
+            {
+                drawable.Fill = SvgExtensions.GetFillSKPaint(svgLine, drawable.TransformedBounds, ignoreAttributes, drawable.Disposable);
+                if (drawable.Fill is null)
+                {
+                    canDrawFill = false;
+                }
+            }
+
+            if (SvgExtensions.IsValidStroke(svgLine, drawable.TransformedBounds))
+            {
+                drawable.Stroke = SvgExtensions.GetStrokeSKPaint(svgLine, drawable.TransformedBounds, ignoreAttributes, drawable.Disposable);
+                if (drawable.Stroke is null)
+                {
+                    canDrawStroke = false;
+                }
+            }
+
+            if (canDrawFill && !canDrawStroke)
+            {
+                drawable.IsDrawable = false;
+                return drawable;
+            }
+
+            SvgExtensions.CreateMarkers(svgLine, drawable.Path, skOwnerBounds, ref drawable.MarkerDrawables, drawable.Disposable);
+
+            // TODO: Transform _skBounds using _skMatrix.
+            drawable.TransformedBounds = drawable.Transform.MapRect(drawable.TransformedBounds);
+
+            return drawable;
+        }
+    }
+}
